[feat]#95 double형으로 변환

+++ app/src/main/java/kr/co/nottodo/view/calendar/weekly/WeeklyViewHolder.kt
@@ -65,7 +65,7 @@ class WeeklyViewHolder(
         }
     }
 
-    fun onNotToDoBind(weeklyDate: LocalDate, notToDoCount: Float) {
+    fun onNotToDoBind(weeklyDate: LocalDate, notToDoCount: Double) {
         val date: Date = Date.valueOf(weeklyDate.toString()) as Date
         this.weeklyDate = weeklyDate
         binding.apply {
